jwt auth dokoncan

+++ src/blog/blog.controller.ts
@@ -1,15 +1,29 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import {
+  Controller,
+  Get,
+  Post,
+  Body,
+  Patch,
+  Param,
+  Delete,
+  UseGuards,
+  Request,
+  BadRequestException
+} from '@nestjs/common';
 import { BlogService } from './blog.service';
 import { CreateBlogDto } from './dto/create-blog.dto';
 import { UpdateBlogDto } from './dto/update-blog.dto';
+import {JwtAuthGuard} from "../auth/guards/jwtAuth.guard";
+import {Blog} from "../entities/blog.entity";
 
-@Controller('blog')
+@Controller('blogs')
 export class BlogController {
   constructor(private readonly blogService: BlogService) {}
 
+  @UseGuards(JwtAuthGuard)
   @Post()
-  create(@Body() createBlogDto: CreateBlogDto) {
-    return this.blogService.create(createBlogDto);
+  create(@Request() req,@Body() createBlogDto: CreateBlogDto) {
+    return this.blogService.create(req.user.id,createBlogDto);
   }
 
   @Get()
@@ -22,13 +36,25 @@ export class BlogController {
     return this.blogService.findOne(+id);
   }
 
+  @UseGuards(JwtAuthGuard)
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateBlogDto: UpdateBlogDto) {
+  async update(@Request() req, @Param('id') id: string, @Body() updateBlogDto: UpdateBlogDto) {
+    const currentUser = req.user.id;
+    const blog: Blog = await this.findOne(id);
+    if (currentUser != blog.user.id) {
+      throw new BadRequestException("It is not your blog");
+    }
     return this.blogService.update(+id, updateBlogDto);
   }
 
+  @UseGuards(JwtAuthGuard)
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  async remove(@Request() req, @Param('id') id: string) {
+    const currentUser = req.user.id;
+    const blog: Blog = await this.blogService.findOne(+id);
+    if (currentUser != blog.user.id) {
+      throw new BadRequestException("It is not your blog");
+    }
     return this.blogService.remove(+id);
   }
 }
